// src/app/Models/Usuario.ts
export class Usuario {
    id: number = 0;
    nombre: string = "";
    apellido: string = "";
    tipo_documento: string = "";
    nro_documento: number = 0;
    correo: string = "";
    username: string = "";
    password: string = "";
    direccion: string = "";
    celular: number = 0;
    rol: string = "";
}